# src/main/webapp/META-INF/stub/swagger.yaml
---
swagger: "2.0"
info:
  title: "Customer management micro service api"
  version: "1.0.0"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /customers:
    get:
      responses:
        '200':
          description: "200 OK"
          schema:
              $ref: "#/definitions/customers"
      operationId: "getCustomers"
    post:
      parameters:
        - name: "customer"
          required: true
          in: "body"
          schema:
            $ref: "#/definitions/customer"
      responses:
        '201':
          description: "Contact created"
      operationId: "postCustomer"
  /customers/{id}:
    parameters:
    - name: "id"
      type: "string"
      required: true
      in: "path"
    get:
      responses:
        200:
          description: "200 OK"
          schema:
            $ref: "#/definitions/customer"
      operationId: "getCustomerById"
    put:
      responses:
        200:
          description: "200 OK"
          schema:
            $ref: "#/definitions/customer"
      operationId: "putCustomer"
      parameters:
      - name: "customer"
        required: true
        in: "body"
        schema:
          $ref: "#/definitions/customer"
    delete:
      responses:
        204:
          description: "204 Entity successfully deleted"
          schema:
            type: "string"
      operationId: "deleteCustomer"
  /customers/email/{email}:
    parameters:
    - name: "email"
      type: "string"
      required: true
      in: "path"
    get:
      responses:
        200:
          description: "200 OK"
          schema:
            $ref: "#/definitions/customer"
      operationId: "getCustomerByEmail"
  /accounts:
     get:
        responses:
          '200':
            description: "200 OK"
            schema:
                $ref: "#/definitions/accounts"
        operationId: "getAccounts"
  /accounts/{accountNumber}:
    parameters:
    - name: "accountNumber"
      type: "string"
      required: true
      in: "path"
    get:
      responses:
        200:
          description: "200 OK"
          schema:
            $ref: "#/definitions/account"
      operationId: "getAccountByNumber"
definitions:
  customer:
    properties:
      id:
        type: "integer"
        description: "unique id"
        format: "int64"
      name:
        type: "string"
        description: "customer's name"
      firstName:
        type: "string"
        description: "customer's first name"
      lastName:
        type: "string"
        description: "customer's last name"
      emailAddress:
        type: "string"
        description: "customer's email address"
      age:
        type: "integer"
        format: "int32"
      gender:
        type: "string"
        description: "customer's gender"
      type:
        type: "string"
        description: "customer's classification or type: Person or Organization"
      status:
        type: "string"
        description: "customer's status: Person or Organization"
      children:
        type: "integer"
        format: "int32"
        description: "Number of dependants"
      estimatedIncome:
        type: "integer"
        format: "double"
        description: "Number of dependants"
      carOwner:
        type: "boolean"
      profession:
        type: "string"
      churn:
        type: "string"
      zipcode:
        type: "string"
      maritalStatus:
        type: "string"
        description: "One of: Married, Single"
      mostDominantTone:
        type: "string"
        description: "Derived attribute, output"
      accountNumber:
        type: "string"
        description: "Customer has one account. It is output from backend"
      longDistance:
        type: "integer"
        format: "double"
      longDistanceBillType:
        type: "string"
      international:
        type: "integer"
        format: "double"
      local:
        type: "integer"
        format: "double"
      balance:
        type: "integer"
        format: "double"
      usage:
        type: "integer"
        format: "double"
      dropped:
        type: "integer"
        format: "int32"
      paymentMethod:
        type: "string"
      localBillType:
        type: "string"
      ratePlan:
        type: "string"
      deviceOwned:
        type: "string"
    additionalProperties: false
  customers:
    type: "array"
    items:
      $ref: "#/definitions/customer"
  account:
    properties:
      id:
        type: "integer"
        description: "unique id"
        format: "int64"
      accountNumber:
        type: "string"
        description: "account number"
      balance:
        type: "integer"
        format: "double"
      longDistance:
        type: "integer"
        format: "double"
      longDistanceBillType:
        type: "string"
      international:
        type: "integer"
        format: "double"
      local:
        type: "integer"
        format: "double"
      usage:
        type: "integer"
        format: "double"
      dropped:
        type: "integer"
        format: "int32"
      paymentMethod:
        type: "string"
      localBillType:
        type: "string"
      ratePlan:
        type: "string"
      deviceOwned:
        type: "string"
  accounts:
    type: "array"
    items:
      $ref: "#/definitions/account"